refactor: cognito session check using a custom hook
